Rename: 폴더 구조 변경
- (section) 폴더 삭제 및 components 폴더로 파일 이동
