+++ front/src/app/profile/[id]/components/ProgressBar.tsx
