+++ front/src/app/profile/[id]/components/Projects.tsx
