+++ front/src/app/profile/[id]/components/ReviewStatistics.tsx
