+++ front/src/app/profile/[id]/components/Studies.tsx
