+++ front/src/app/profile/[id]/components/UserInfo.tsx
